+++ mmengine/_strategy/__init__.py
@@ -1,8 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .base_strategy import BaseStrategy
-from .ddp_strategy import DDPStrategy
-from .deepspeed_strategy import DeepSpeedStrategy
-from .single_device_strategy import SingleDeviceStrategy
+from .base import BaseStrategy
+from .deepspeed import DeepSpeedStrategy
+from .distributed import DDPStrategy
+from .single_device import SingleDeviceStrategy
 
 __all__ = [
     'BaseStrategy', 'DDPStrategy', 'SingleDeviceStrategy', 'DeepSpeedStrategy'
